Add files via upload
Open data source website: https://www.kaggle.com/vjchoudhary7/hr-analytics-case-study
Problem StatementA large company named Ham-3, employs, at any given point of time, around 4000 employees. However, every year, around 15% of its employees leave the company and need to be replaced with the talent pool available in the job market. The management believes that this level of attrition (employees leaving, either on their own or because they got fired) is bad for the company, because for the following reasons -
The former employees’ projects get delayed, which makes it difficult to meet timelines, resulting in a reputation loss among consumers and partners
A sizeable department has to be maintained, for the purposes of recruiting new talent
More often than not, the new employees have to be trained for the job and/or given time to acclimatize themselves to the company

In conclusion alt I hypothesis that Income was the main contributing factor to the company’s high attrition rate. 
Null hypothesis- income has no relationship with staff leaving the company. 
Our Initial Hypothesis of monthly income and attrition correlation was proved false by our model and we found Years with the current Managers to be the main cause of Staff turnover. Years with the current manager was the main cause of employee attrition. We Recommend:
Allocating Fund for Training Managers on how to better handle Employees.
Employee Bonding Trips and Gatherings
Focus on Career Development of Staff and increased Stock options after working with the company after 5 years.